refactor(engine): move StreamProcessorMode to streamprocessor package
